requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
jinja2==3.1.3
python-multipart==0.0.9
transformers==4.43.3
accelerate==0.34.2
torch==2.6.0
torchvision==0.21.0
pillow==10.4.0


